+++ CHANGELOG.md
@@ -87,6 +87,8 @@ ALTER EXTENSION pg_clickhouse UPDATE TO '0.11';
 
 ### 🐞 Bug Fixes
 
+*   Fixed inserting `interval` values with HTTP driver. Values are now
+    encoded based on destination's ClickHouse schema ([#376]).
 *   Fixed an issue where a check for ordered aggregates incorrectly handled
     custom aggregates that were not part of an extension. It now properly
     keeps custom aggregate execution local. Thanks to Minh Vu for the PR
@@ -129,6 +131,8 @@ ALTER EXTENSION pg_clickhouse UPDATE TO '0.11';
     "ClickHouse/pg_clickhouse#367 CH 26.9 map cardinality to arrayFlattenedLength"
   [#374]: https://github.com/ClickHouse/pg_clickhouse/pull/374
     "ClickHouse/pg_clickhouse#374 update pg-clickhouse-c, improves integer/interval coercion"
+  [#376]: https://github.com/ClickHouse/pg_clickhouse/pull/376
+    "ClickHouse/pg_clickhouse#376 http: DESCRIBE TABLE to know destination schema"
   [arrayFlattenedLength]: https://clickhouse.com/docs/reference/functions/regular-functions/array-functions#arrayFlattenedLength
     "ClickHouse Docs: arrayFlattenedLength"
   [#373]: https://github.com/ClickHouse/pg_clickhouse/pull/373